1. Edited the RegisterActivity with the same layout as the DashActivity with some few TextViews.
2. Also enabled registration for new members on the RegisterActivity, so before you access the DashActivity, you need to be logged in.

3. Also added a back button to the RegisterActivity, so once you are registered, you can go back to the MainActivity

3. Changed names of LoginActivity.java to MainActivity

4. Fixed bug preventing MainActivity to go into DashActivity, deleted the old xml and created a new, we are still using cardViews.

5. created 6 cardViews in the DashActivity.
