--- app/src/main/java/com/egcoding/acmapp/LoginActivity.java
@@ -1,55 +1,0 @@
-package com.egcoding.acmapp;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-
-
-public class LoginActivity extends AppCompatActivity {
-    private EditText email;
-    private EditText password;
-    private Button login;
-
-    private FirebaseAuth auth;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
-
-        email=findViewById(R.id.email);
-        password=findViewById(R.id.password);
-        login=findViewById(R.id.login);
-
-        auth = FirebaseAuth.getInstance();
-        login.setOnClickListener(new View.OnClickListener() { // whenever login button is clicked
-            @Override
-            public void onClick(View view) {
-                String txt_email = email.getText().toString(); // grabbing the inputed email and converting to string variable
-                String txt_password = password.getText().toString();
-
-                loginUser(txt_email,txt_password);
-            }
-        });
-    }
-
-    private void loginUser(String email, String password) {
-        auth.signInWithEmailAndPassword(email,password).addOnSuccessListener(new OnSuccessListener<AuthResult>() {
-            @Override
-            public void onSuccess(AuthResult authResult) {
-                Toast.makeText(LoginActivity.this, "Login Successful!", Toast.LENGTH_SHORT).show();
-                startActivity(new Intent(LoginActivity.this,MainActivity.class));
-                finish();
-            }
-        });
-    }
-}
