+++ app/src/main/res/drawable/gradientdarkgray.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient android:angle="45"
+        android:startColor="@color/lightergray"
+        android:centerColor="@color/gray"
+        android:endColor="@color/darkgray"/>
+
+</shape>
