+++ app/src/main/res/drawable/ic_launcher_background.xml
@@ -5,7 +5,7 @@
     android:viewportWidth="108"
     android:viewportHeight="108">
     <path
-        android:fillColor="#3DDC84"
+        android:fillColor="#008577"
         android:pathData="M0,0h108v108h-108z" />
     <path
         android:fillColor="#00000000"
