+++ app/src/main/res/drawable/login_rounded_button.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    tools:ignore="ExtraText">
+
+    android:shape ="rectangle">
+    <solid android:color ="@color/red"/>
+    <corners android:radius = "75dp"/>
+
+</shape>
