+++ app/src/main/res/drawable/register_rounded_button.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color ="@color/white"/>
+    <corners android:radius="75dp"/>
+
+</shape>
