+++ app/src/main/res/drawable/white_to_red_highlight.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:state_pressed="true" android:drawable="@drawable/login_rounded_button" />
+    <item android:drawable="@drawable/register_rounded_button" />
+</selector>
