+++ app/src/main/res/layout/activity_register.xml
@@ -1,33 +1,104 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:background="@drawable/gradientdarkgray"
     tools:context="com.egcoding.acmapp.RegisterActivity">
 
-    <EditText
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbar"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="email"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
+
+    <ImageView
+        android:id="@+id/whiteAcmLogo"
+        android:layout_width="200dp"
+        android:layout_height="150dp"
+        android:layout_gravity="center"
+        android:layout_marginTop="50dp"
+        android:background="@drawable/acmwhitelogo" />
+
+
+    <EditText
         android:id="@+id/email"
-        android:layout_marginTop="200dp"/>
+        android:layout_width="275dp"
+        android:layout_height="40dp"
+        android:layout_marginLeft="50dp"
+        android:layout_marginTop="60dp"
+        android:background="@android:drawable/screen_background_light_transparent"
+        android:ems="10"
+        android:fontFamily="sans-serif-condensed-medium"
+        android:hint="@string/email"
+        android:inputType="textEmailAddress"
+        android:paddingLeft="3dp"
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        android:textStyle="bold"
+        android:layout_marginStart="50dp" />
+    <EditText
+        android:id="@+id/password"
+        android:layout_width="275dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginLeft="50dp"
+        android:background="@android:drawable/screen_background_light_transparent"
+        android:ems="10"
+        android:fontFamily="sans-serif-condensed-medium"
+        android:hint="@string/password"
+        android:inputType="textPassword"
+        android:paddingLeft="3dp"
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        android:layout_marginStart="50dp" />
 
     <EditText
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="password"
-        android:layout_marginTop="20dp"
-        android:layout_below="@id/email"
-        android:id="@+id/password"/>
+        android:id="@+id/fullname"
+        android:layout_width="275dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginLeft="50dp"
+        android:background="@android:drawable/screen_background_light_transparent"
+        android:ems="10"
+        android:fontFamily="sans-serif-condensed-medium"
+        android:hint="@string/my_man_what_s_your_name"
+        android:inputType="text"
+        android:paddingLeft="3dp"
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        android:layout_marginStart="50dp" />
 
+    <EditText
+        android:id="@+id/username"
+        android:layout_width="275dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginLeft="50dp"
+        android:background="@android:drawable/screen_background_light_transparent"
+        android:ems="10"
+        android:fontFamily="sans-serif-condensed-medium"
+        android:hint="@string/username"
+        android:inputType="text"
+        android:paddingLeft="3dp"
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        android:layout_marginStart="50dp" />
 
     <Button
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_width="250dp"
+        android:layout_height="40dp"
         android:layout_marginTop="20dp"
-        android:layout_below="@id/password"
-        android:text="register"
-        android:id="@+id/register"/>
+        android:id="@+id/register"
+        android:layout_marginLeft="65dp"
+        android:background="@drawable/red_to_gray_highlight"
+        android:text="Register"
+        android:textColor="@color/white"
+        android:textSize="14sp"
+        android:textStyle="bold"
+        android:layout_marginStart="65dp" />
 
-</RelativeLayout>
+</LinearLayout>
